chore(lambda): refactors for testability.
